Duplicate SSAFlags in SSAOpts and move them back to cmd/tools
