+++ pkg/kapp/cmd/app/delete.go
@@ -9,7 +9,6 @@ import (
 	ctlcap "github.com/k14s/kapp/pkg/kapp/clusterapply"
 	cmdcore "github.com/k14s/kapp/pkg/kapp/cmd/core"
 	cmdtools "github.com/k14s/kapp/pkg/kapp/cmd/tools"
-	"github.com/k14s/kapp/pkg/kapp/cmd/tools/ssa"
 	ctlconf "github.com/k14s/kapp/pkg/kapp/config"
 	ctldiff "github.com/k14s/kapp/pkg/kapp/diff"
 	ctldgraph "github.com/k14s/kapp/pkg/kapp/diffgraph"
@@ -64,7 +63,7 @@ func (o *DeleteOptions) Run() error {
 	// which passes field manager name to K8S API.
 	// This name is not persisted anywhere in managedFields and it's value doesn't really matter, therefore there
 	// is no need to expose --ssa-field-manager CLI flag in the delete command. Set it to something reasonable.
-	ssaFlags := ssa.SSAFlags{
+	ssaFlags := cmdtools.SSAFlags{
 		FieldManagerName: "kapp-shouldnt-be-seen-anywhere",
 	}
 	app, supportObjs, err := Factory(o.depsFactory, o.AppFlags, o.ResourceTypesFlags, o.logger, &ssaFlags)
