+++ pkg/kapp/cmd/app/factory.go
@@ -6,7 +6,7 @@ package app
 import (
 	ctlapp "github.com/k14s/kapp/pkg/kapp/app"
 	cmdcore "github.com/k14s/kapp/pkg/kapp/cmd/core"
-	"github.com/k14s/kapp/pkg/kapp/cmd/tools/ssa"
+	"github.com/k14s/kapp/pkg/kapp/cmd/tools"
 	"github.com/k14s/kapp/pkg/kapp/logger"
 	ctlres "github.com/k14s/kapp/pkg/kapp/resources"
 	"k8s.io/client-go/kubernetes"
@@ -21,7 +21,7 @@ type FactorySupportObjs struct {
 }
 
 func FactoryClients(depsFactory cmdcore.DepsFactory, nsFlags cmdcore.NamespaceFlags,
-	resTypesFlags ResourceTypesFlags, logger logger.Logger, ssaFlags *ssa.SSAFlags) (FactorySupportObjs, error) {
+	resTypesFlags ResourceTypesFlags, logger logger.Logger, ssaFlags *tools.SSAFlags) (FactorySupportObjs, error) {
 
 	coreClient, err := depsFactory.CoreClient()
 	if err != nil {
@@ -46,7 +46,11 @@ func FactoryClients(depsFactory cmdcore.DepsFactory, nsFlags cmdcore.NamespaceFl
 	resourcesImplOpts := ctlres.ResourcesImplOpts{
 		FallbackAllowedNamespaces:        []string{nsFlags.Name},
 		ScopeToFallbackAllowedNamespaces: resTypesFlags.ScopeToFallbackAllowedNamespaces,
-		SSAFlags:                         ssaFlags,
+		SSA: &ctlres.SSAOpts{
+			Enabled:          ssaFlags.Enabled,
+			ForceConflict:    ssaFlags.ForceConflict,
+			FieldManagerName: ssaFlags.FieldManagerName,
+		},
 	}
 
 	resources := ctlres.NewResourcesImpl(
@@ -67,7 +71,7 @@ func FactoryClients(depsFactory cmdcore.DepsFactory, nsFlags cmdcore.NamespaceFl
 }
 
 func Factory(depsFactory cmdcore.DepsFactory, appFlags Flags,
-	resTypesFlags ResourceTypesFlags, logger logger.Logger, ssaFlags *ssa.SSAFlags) (ctlapp.App, FactorySupportObjs, error) {
+	resTypesFlags ResourceTypesFlags, logger logger.Logger, ssaFlags *tools.SSAFlags) (ctlapp.App, FactorySupportObjs, error) {
 
 	supportingObjs, err := FactoryClients(depsFactory, appFlags.NamespaceFlags, resTypesFlags, logger, ssaFlags)
 	if err != nil {
