+++ pkg/kapp/cmd/tools/diff_flags.go
@@ -6,7 +6,6 @@ package tools
 import (
 	"fmt"
 	ctlcap "github.com/k14s/kapp/pkg/kapp/clusterapply"
-	"github.com/k14s/kapp/pkg/kapp/cmd/tools/ssa"
 	ctldiff "github.com/k14s/kapp/pkg/kapp/diff"
 	"github.com/spf13/cobra"
 )
@@ -46,7 +45,7 @@ func (s *DiffFlags) SetWithPrefix(prefix string, cmd *cobra.Command) {
 	cmd.Flags().StringVar(&s.Filter, prefix+"filter", "", `Set changes filter (example: {"and":[{"ops":["update"]},{"existingResource":{"kinds":["Deployment"]}]})`)
 }
 
-func AdjustDiffFlags(ssa ssa.SSAFlags, df *DiffFlags, diffPrefix string, cmd *cobra.Command) error {
+func AdjustDiffFlags(ssa SSAFlags, df *DiffFlags, diffPrefix string, cmd *cobra.Command) error {
 	if len(diffPrefix) > 0 {
 		diffPrefix = diffPrefix + "-"
 	}
